+++ crest_data_systems_netapp_aiqum/CHANGELOG.md
@@ -1,4 +1,9 @@
 # CHANGELOG - crest_data_systems_netapp_aiqum
 
+## 1.0.1 / 2023-04-18
+
+- [Updated] Support and setup information in readme.
+
 ## 1.0.0 / 2023-01-25
-- [Added] Initial release.
+
+- [Added] NetApp AIQUM version <=`NetApp Release  9.12RC1` is supported in this release.